requirements.txt
Django==2.1
djangorestframework==3.8.2
django-filter==2.0.0
coreapi==2.3.3
Pillow==5.2.0
python-magic==0.4.15